Improved spec of ISO8601
