+++ kazuki-toolkit/src/test/kotlin/com/anaplan/engineering/kazuki/toolkit/iso8601/TestIso8601.kt
@@ -1,24 +1,40 @@
-package com.anaplan.engineering.kazuki.toolkit
+package com.anaplan.engineering.kazuki.toolkit.iso8601
 
 import com.anaplan.engineering.kazuki.core.InvariantFailure
 import com.anaplan.engineering.kazuki.core.PreconditionFailure
 import com.anaplan.engineering.kazuki.core.mk_Seq
 import com.anaplan.engineering.kazuki.core.mk_Set1
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.*
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Date_Module.mk_Date
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.DtgInZone_Module.mk_DtgInZone
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Dtg_Module.mk_Dtg
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Duration_Module.mk_Duration
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Interval_Module.mk_Interval
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.NormalisedTime_Module.mk_NormalisedTime
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Offset_Module.mk_Offset
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.TimeInZone_Module.mk_TimeInZone
-import com.anaplan.engineering.kazuki.toolkit.ISO8601.Time_Module.mk_Time
-import junit.framework.TestCase.assertEquals
-import kotlin.test.Test
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Date_Module.mk_Date
+import com.anaplan.engineering.kazuki.toolkit.iso8601.DtgInZone_Module.mk_DtgInZone
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Dtg_Module.mk_Dtg
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Duration_Module.mk_Duration
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Interval_Module.mk_Interval
+import com.anaplan.engineering.kazuki.toolkit.iso8601.NormalisedTime_Module.mk_NormalisedTime
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Offset_Module.mk_Offset
+import com.anaplan.engineering.kazuki.toolkit.iso8601.TimeInZone_Module.mk_TimeInZone
+import com.anaplan.engineering.kazuki.toolkit.iso8601.Time_Module.mk_Time
 import kotlin.test.assertFailsWith
+import kotlin.test.assertEquals
+import kotlin.test.Test
 
-class ISO8601Test {
+class TestIso8601 {
+
+    @Test
+    fun constructDuration() {
+        assertEquals(10L, mk_Duration(10).milliseconds)
+    }
+
+    @Test
+    fun t1() {
+        DtgUtilities.dtgDiff(
+            mk_Dtg(mk_Date(2018, 4, 1), mk_Time(0, 0, 0, 10)),
+            mk_Dtg(mk_Date(2018, 4, 1), mk_Time(0, 0, 0, 0)),
+        )
+        DtgUtilities.dtgDiff(
+            mk_Dtg(mk_Date(2018, 4, 1), mk_Time(10, 0, 0, 0)),
+            mk_Dtg(mk_Date(2018, 4, 1), mk_Time(0, 0, 0, 0)),
+        )
+    }
 
     @Test
     fun timeInvariantTest() {
@@ -47,10 +63,22 @@ class ISO8601Test {
     @Test
     fun dtgInZoneInvariantTest() {
         assertFailsWith<InvariantFailure> {
-            mk_DtgInZone(FirstDate, mk_TimeInZone(FirstTime, mk_Offset(Duration.fromHours(1), PlusOrMinus.Plus)))
+            mk_DtgInZone(
+                FirstDate,
+                mk_TimeInZone(
+                    FirstTime,
+                    mk_Offset(Duration.fromHours(1), PlusOrMinus.Plus)
+                )
+            )
         }
         assertFailsWith<InvariantFailure> {
-            mk_DtgInZone(LastDate, mk_TimeInZone(LastTime, mk_Offset(Duration.fromHours(1), PlusOrMinus.Minus)))
+            mk_DtgInZone(
+                LastDate,
+                mk_TimeInZone(
+                    LastTime,
+                    mk_Offset(Duration.fromHours(1), PlusOrMinus.Minus)
+                )
+            )
         }
     }
 
@@ -67,52 +95,52 @@ class ISO8601Test {
 
     @Test
     fun isLeapTest() {
-        assertEquals(true, isLeap(1992))
-        assertEquals(false, isLeap(1993))
+        assertEquals(true, DateUtilities.isLeap(1992))
+        assertEquals(false, DateUtilities.isLeap(1993))
     }
 
     @Test
     fun daysInMonthTest() {
-        assertEquals(29, daysInMonth(1992, 2))
-        assertEquals(28, daysInMonth(1991, 2))
-        assertEquals(30, daysInMonth(1997, 9))
+        assertEquals(29, DateUtilities.daysInMonth(1992, 2))
+        assertEquals(28, DateUtilities.daysInMonth(1991, 2))
+        assertEquals(30, DateUtilities.daysInMonth(1997, 9))
     }
 
     @Test
     fun daysInYearTest() {
-        assertEquals(366, daysInYear(1992))
-        assertEquals(365, daysInYear(1))
-        assertEquals(365, daysInYear(1991))
+        assertEquals(366, DateUtilities.daysInYear(1992))
+        assertEquals(365, DateUtilities.daysInYear(1))
+        assertEquals(365, DateUtilities.daysInYear(1991))
     }
 
     @Test
     fun dtgInRangeDayTest() {
         assertEquals(
             true,
-            mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay().functions.inRange(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 3).properties.dtgAtStartOfDay.functions.inRange(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             false,
-            mk_Date(1990, 1, 7).functions.toDtgAtStartOfDay().functions.inRange(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 7).properties.dtgAtStartOfDay.functions.inRange(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             true,
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.inRange(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.inRange(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             false,
-            mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay().functions.inRange(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 3).properties.dtgAtStartOfDay.functions.inRange(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
             )
         )
     }
@@ -154,31 +182,31 @@ class ISO8601Test {
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
-            ).functions.contains(mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay())
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
+            ).functions.contains(mk_Date(1990, 1, 3).properties.dtgAtStartOfDay)
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
-            ).functions.contains(mk_Date(1990, 1, 7).functions.toDtgAtStartOfDay())
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
+            ).functions.contains(mk_Date(1990, 1, 7).properties.dtgAtStartOfDay)
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
-            ).functions.contains(mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay())
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
+            ).functions.contains(mk_Date(1990, 1, 1).properties.dtgAtStartOfDay)
 
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
-            ).functions.contains(mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay())
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
+            ).functions.contains(mk_Date(1990, 1, 3).properties.dtgAtStartOfDay)
         )
         assertEquals(
             true,
@@ -214,32 +242,32 @@ class ISO8601Test {
     fun dtgWithinDurationOfDtgTest() {
         assertEquals(
             true,
-            mk_Date(1989, 1, 3).functions.toDtgAtStartOfDay().functions.withinDurationOfDtg(
-                Duration.fromDays(3), mk_Date(1989, 1, 1).functions.toDtgAtStartOfDay()
+            mk_Date(1989, 1, 3).properties.dtgAtStartOfDay.functions.withinDurationOfDtg(
+                Duration.fromDays(3), mk_Date(1989, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             true,
-            mk_Date(1989, 12, 30).functions.toDtgAtStartOfDay().functions.withinDurationOfDtg(
-                Duration.fromDays(3), mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            mk_Date(1989, 12, 30).properties.dtgAtStartOfDay.functions.withinDurationOfDtg(
+                Duration.fromDays(3), mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             true,
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.withinDurationOfDtg(
-                Duration.fromDays(0), mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.withinDurationOfDtg(
+                Duration.fromDays(0), mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             false,
-            mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay().functions.withinDurationOfDtg(
-                Duration.fromDays(3), mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            mk_Date(1990, 1, 6).properties.dtgAtStartOfDay.functions.withinDurationOfDtg(
+                Duration.fromDays(3), mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             false,
-            mk_Date(1989, 12, 27).functions.toDtgAtStartOfDay().functions.withinDurationOfDtg(
-                Duration.fromDays(3), mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            mk_Date(1989, 12, 27).properties.dtgAtStartOfDay.functions.withinDurationOfDtg(
+                Duration.fromDays(3), mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
     }
@@ -248,38 +276,38 @@ class ISO8601Test {
     fun dtgInIntervalTest() {
         assertEquals(
             true,
-            mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay().functions.inInterval(
+            mk_Date(1990, 1, 3).properties.dtgAtStartOfDay.functions.inInterval(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
-            mk_Date(1990, 1, 7).functions.toDtgAtStartOfDay().functions.inInterval(
+            mk_Date(1990, 1, 7).properties.dtgAtStartOfDay.functions.inInterval(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.inInterval(
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.inInterval(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
 
         )
         assertEquals(
             false,
-            mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay().functions.inInterval(
+            mk_Date(1990, 1, 3).properties.dtgAtStartOfDay.functions.inInterval(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
                 )
             )
         )
@@ -328,72 +356,72 @@ class ISO8601Test {
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 4).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 4).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 4).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 4).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 4).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 4).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 8).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 8).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.overlap(
                 mk_Interval(
-                    mk_Date(1990, 1, 8).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 8).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 10).properties.dtgAtStartOfDay
                 )
             )
         )
@@ -404,96 +432,96 @@ class ISO8601Test {
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 10).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 12).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 14).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 12).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 14).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 10).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 5).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 10).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 8).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 12).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 8).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 12).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 5).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 10).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             false,
             mk_Interval(
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 3).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 3).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 3).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
         assertEquals(
             true,
             mk_Interval(
-                mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 5).properties.dtgAtStartOfDay
             ).functions.within(
                 mk_Interval(
-                    mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-                    mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
+                    mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+                    mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
                 )
             )
         )
@@ -502,12 +530,12 @@ class ISO8601Test {
     @Test
     fun dtgAddDurationTest() {
         assertEquals(
-            mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay().functions.addDuration(Duration.fromDays(3))
+            mk_Date(1990, 1, 5).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 2).properties.dtgAtStartOfDay.functions.addDuration(Duration.fromDays(3))
         )
         assertEquals(
-            mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay().functions.addDuration(Duration.fromDays(0))
+            mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 2).properties.dtgAtStartOfDay.functions.addDuration(Duration.fromDays(0))
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 2), mk_Time(5, 20, 0, 0)),
@@ -535,28 +563,37 @@ class ISO8601Test {
     fun dtgSubtractDurationTest() {
         assertFailsWith<PreconditionFailure> { Duration.fromSeconds(1).functions.subtractDuration(Duration.fromSeconds(5)) }
         assertEquals(
-            mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay().functions.subtractDuration(Duration.fromDays(3))
+            mk_Date(1990, 1, 2).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 5).properties.dtgAtStartOfDay.functions.subtractDuration(Duration.fromDays(3))
         )
         assertEquals(
-            mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 5).functions.toDtgAtStartOfDay().functions.subtractDuration(Duration.fromDays(0))
+            mk_Date(1990, 1, 5).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 5).properties.dtgAtStartOfDay.functions.subtractDuration(Duration.fromDays(0))
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 5), mk_Time(3, 20, 0, 0)),
-            mk_Dtg(mk_Date(1990, 1, 5), mk_Time(6, 40, 0, 0)).functions.subtractDuration(
+            mk_Dtg(
+                mk_Date(1990, 1, 5),
+                mk_Time(6, 40, 0, 0)
+            ).functions.subtractDuration(
                 Duration.fromHours(3).functions.addDuration(Duration.fromMinutes(20))
             )
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 5), mk_Time(2, 40, 0, 0)),
-            mk_Dtg(mk_Date(1990, 1, 5), mk_Time(6, 20, 0, 0)).functions.subtractDuration(
+            mk_Dtg(
+                mk_Date(1990, 1, 5),
+                mk_Time(6, 20, 0, 0)
+            ).functions.subtractDuration(
                 Duration.fromHours(3).functions.addDuration(Duration.fromMinutes(40))
             )
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 2), mk_Time(2, 20, 20, 5)),
-            mk_Dtg(mk_Date(1990, 1, 6), mk_Time(5, 40, 30, 10)).functions.subtractDuration(
+            mk_Dtg(
+                mk_Date(1990, 1, 6),
+                mk_Time(5, 40, 30, 10)
+            ).functions.subtractDuration(
                 Duration.fromDays(4).functions.addDuration(
                     Duration.fromHours(3).functions.addDuration(
                         Duration.fromMinutes(20).functions.addDuration(
@@ -582,7 +619,7 @@ class ISO8601Test {
                     )
                 )
             ),
-            dtgDiff(
+            DtgUtilities.dtgDiff(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 1, 1, 1)),
                 mk_Dtg(mk_Date(1990, 1, 6), mk_Time(6, 6, 6, 6))
             )
@@ -597,14 +634,14 @@ class ISO8601Test {
                     )
                 )
             ),
-            dtgDiff(
+            DtgUtilities.dtgDiff(
                 mk_Dtg(mk_Date(1990, 1, 6), mk_Time(6, 6, 6, 6)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 1, 1, 1))
             )
         )
         assertEquals(
             Duration.fromDays(0),
-            dtgDiff(
+            DtgUtilities.dtgDiff(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 1, 1, 1)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 1, 1, 1))
             )
@@ -619,7 +656,7 @@ class ISO8601Test {
                     )
                 )
             ),
-            dtgDiff(
+            DtgUtilities.dtgDiff(
                 mk_Dtg(mk_Date(1990, 1, 31), mk_Time(1, 1, 1, 1)),
                 mk_Dtg(mk_Date(1990, 2, 5), mk_Time(6, 6, 6, 6))
             )
@@ -634,7 +671,7 @@ class ISO8601Test {
                     )
                 )
             ),
-            dtgDiff(
+            DtgUtilities.dtgDiff(
                 mk_Dtg(mk_Date(1990, 2, 5), mk_Time(6, 6, 6, 6)),
                 mk_Dtg(mk_Date(1990, 1, 31), mk_Time(1, 1, 1, 1))
             )
@@ -651,10 +688,19 @@ class ISO8601Test {
 
     @Test
     fun durationSubtractTest() {
-        assertEquals(Duration.fromDays(5), Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(3)))
+        assertEquals(
+            Duration.fromDays(5),
+            Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(3))
+        )
         assertFailsWith<PreconditionFailure> { Duration.fromDays(2).functions.subtractDuration(Duration.fromDays(3)) }
-        assertEquals(Duration.fromDays(8), Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(0)))
-        assertEquals(Duration.fromDays(0), Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(8)))
+        assertEquals(
+            Duration.fromDays(8),
+            Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(0))
+        )
+        assertEquals(
+            Duration.fromDays(0),
+            Duration.fromDays(8).functions.subtractDuration(Duration.fromDays(8))
+        )
     }
 
     @Test
@@ -675,19 +721,37 @@ class ISO8601Test {
 
     @Test
     fun durationDiffTest() {
-        assertEquals(Duration.fromDays(5), durationDiff(Duration.fromDays(10), Duration.fromDays(5)))
-        assertEquals(Duration.fromDays(5), durationDiff(Duration.fromDays(0), Duration.fromDays(5)))
-        assertEquals(Duration.fromDays(5), durationDiff(Duration.fromDays(5), Duration.fromDays(10)))
-        assertEquals(Duration.fromDays(0), durationDiff(Duration.fromDays(10), Duration.fromDays(10)))
-        assertEquals(Duration.fromDays(1999999995), durationDiff(Duration.fromDays(2000000000), Duration.fromDays(5)))
-        assertEquals(Duration.fromMillis(15), durationDiff(Duration.fromMillis(20), Duration.fromMillis(5)))
+        assertEquals(
+            Duration.fromDays(5),
+            DurationUtiltites.durationDiff(Duration.fromDays(10), Duration.fromDays(5))
+        )
+        assertEquals(
+            Duration.fromDays(5),
+            DurationUtiltites.durationDiff(Duration.fromDays(0), Duration.fromDays(5))
+        )
+        assertEquals(
+            Duration.fromDays(5),
+            DurationUtiltites.durationDiff(Duration.fromDays(5), Duration.fromDays(10))
+        )
+        assertEquals(
+            Duration.fromDays(0),
+            DurationUtiltites.durationDiff(Duration.fromDays(10), Duration.fromDays(10))
+        )
+        assertEquals(
+            Duration.fromDays(1999999995),
+            DurationUtiltites.durationDiff(Duration.fromDays(2000000000), Duration.fromDays(5))
+        )
+        assertEquals(
+            Duration.fromMillis(15),
+            DurationUtiltites.durationDiff(Duration.fromMillis(20), Duration.fromMillis(5))
+        )
     }
 
     @Test
     fun durationToMillisTest() {
-        assertEquals(12, mk_Duration(12).functions.toMillis())
-        assertEquals(0, mk_Duration(0).functions.toMillis())
-        assertEquals(2000000000, mk_Duration(2000000000).functions.toMillis())
+        assertEquals(12, mk_Duration(12).milliseconds)
+        assertEquals(0, mk_Duration(0).milliseconds)
+        assertEquals(2000000000, mk_Duration(2000000000).milliseconds)
     }
 
     @Test
@@ -699,9 +763,9 @@ class ISO8601Test {
 
     @Test
     fun durationToSecondsTest() {
-        assertEquals(12, mk_Duration(12000).functions.toSeconds())
-        assertEquals(0, mk_Duration(0).functions.toSeconds())
-        assertEquals(2000000, mk_Duration(2000000000).functions.toSeconds())
+        assertEquals(12, mk_Duration(12000).properties.seconds)
+        assertEquals(0, mk_Duration(0).properties.seconds)
+        assertEquals(2000000, mk_Duration(2000000000).properties.seconds)
     }
 
     @Test
@@ -713,9 +777,9 @@ class ISO8601Test {
 
     @Test
     fun durationToMinutesTest() {
-        assertEquals(10, mk_Duration(600000).functions.toMinutes())
-        assertEquals(0, mk_Duration(0).functions.toMinutes())
-        assertEquals(2000, mk_Duration(120000000).functions.toMinutes())
+        assertEquals(10, mk_Duration(600000).properties.minutes)
+        assertEquals(0, mk_Duration(0).properties.minutes)
+        assertEquals(2000, mk_Duration(120000000).properties.minutes)
     }
 
     @Test
@@ -768,10 +832,10 @@ class ISO8601Test {
 
     @Test
     fun durationToHoursTest() {
-        assertEquals(10, mk_Duration(36000000).functions.toHours())
-        assertEquals(0, mk_Duration(0).functions.toHours())
-        assertEquals(200, mk_Duration(720000000).functions.toHours())
-        assertEquals(876600, Duration.durationFromFirstYearUpToStartOfYear(100).functions.toHours())
+        assertEquals(10, mk_Duration(36000000).properties.hours)
+        assertEquals(0, mk_Duration(0).properties.hours)
+        assertEquals(200, mk_Duration(720000000).properties.hours)
+        assertEquals(876600, Duration.durationFromFirstYearUpToStartOfYear(100).properties.hours)
     }
 
     @Test
@@ -803,10 +867,10 @@ class ISO8601Test {
 
     @Test
     fun durationToDaysTest() {
-        assertEquals(10, mk_Duration(864000000).functions.toDays())
-        assertEquals(9, mk_Duration(863999999).functions.toDays())
-        assertEquals(0, mk_Duration(0).functions.toDays())
-        assertEquals(20, mk_Duration(1728000000).functions.toDays())
+        assertEquals(10, mk_Duration(864000000).properties.days)
+        assertEquals(9, mk_Duration(863999999).properties.days)
+        assertEquals(0, mk_Duration(0).properties.days)
+        assertEquals(20, mk_Duration(1728000000).properties.days)
     }
 
     @Test
@@ -818,21 +882,36 @@ class ISO8601Test {
 
     @Test
     fun dateToDayOfWeekTest() {
-        assertEquals(DayOfWeek.Thursday, mk_Date(2009, 8, 13).functions.toDayOfWeek())
-        assertEquals(DayOfWeek.Saturday, mk_Date(2000, 4, 1).functions.toDayOfWeek())
-        assertEquals(DayOfWeek.Monday, mk_Date(1, 1, 1).functions.toDayOfWeek())
-        assertEquals(DayOfWeek.Wednesday, mk_Date(2019, 10, 9).functions.toDayOfWeek())
-        assertEquals(DayOfWeek.Saturday, mk_Date(2017, 10, 28).functions.toDayOfWeek())
-        assertEquals(DayOfWeek.Wednesday, mk_Date(2020, 1, 1).functions.toDayOfWeek())
+        assertEquals(DayOfWeek.Thursday, mk_Date(2009, 8, 13).properties.dayOfWeek)
+        assertEquals(DayOfWeek.Saturday, mk_Date(2000, 4, 1).properties.dayOfWeek)
+        assertEquals(DayOfWeek.Monday, mk_Date(1, 1, 1).properties.dayOfWeek)
+        assertEquals(DayOfWeek.Wednesday, mk_Date(2019, 10, 9).properties.dayOfWeek)
+        assertEquals(DayOfWeek.Saturday, mk_Date(2017, 10, 28).properties.dayOfWeek)
+        assertEquals(DayOfWeek.Wednesday, mk_Date(2020, 1, 1).properties.dayOfWeek)
     }
 
     @Test
     fun dateToDtgTest() {
-        assertEquals(mk_Dtg(mk_Date(1, 1, 1), FirstTime), mk_Date(1, 1, 1).functions.toDtgAtStartOfDay())
-        assertEquals(mk_Dtg(mk_Date(1000, 12, 11), FirstTime), mk_Date(1000, 12, 11).functions.toDtgAtStartOfDay())
-        assertEquals(mk_Dtg(mk_Date(2000, 2, 29), FirstTime), mk_Date(2000, 2, 29).functions.toDtgAtStartOfDay())
-        assertEquals(mk_Dtg(mk_Date(9999, 12, 31), FirstTime), mk_Date(9999, 12, 31).functions.toDtgAtStartOfDay())
-        assertEquals(mk_Dtg(mk_Date(2001, 2, 28), FirstTime), mk_Date(2001, 2, 28).functions.toDtgAtStartOfDay())
+        assertEquals(
+            mk_Dtg(mk_Date(1, 1, 1), FirstTime),
+            mk_Date(1, 1, 1).properties.dtgAtStartOfDay
+        )
+        assertEquals(
+            mk_Dtg(mk_Date(1000, 12, 11), FirstTime),
+            mk_Date(1000, 12, 11).properties.dtgAtStartOfDay
+        )
+        assertEquals(
+            mk_Dtg(mk_Date(2000, 2, 29), FirstTime),
+            mk_Date(2000, 2, 29).properties.dtgAtStartOfDay
+        )
+        assertEquals(
+            mk_Dtg(mk_Date(9999, 12, 31), FirstTime),
+            mk_Date(9999, 12, 31).properties.dtgAtStartOfDay
+        )
+        assertEquals(
+            mk_Dtg(mk_Date(2001, 2, 28), FirstTime),
+            mk_Date(2001, 2, 28).properties.dtgAtStartOfDay
+        )
     }
 
     @Test
@@ -911,15 +990,15 @@ class ISO8601Test {
     @Test
     fun durationToDtgAfterFirstDtgTest() {
         assertEquals(
-            mk_Date(0, 1, 6).functions.toDtgAtStartOfDay(),
+            mk_Date(0, 1, 6).properties.dtgAtStartOfDay,
             Duration.fromDays(5).functions.toDtgAfterFirstDtg()
         )
         assertEquals(
-            mk_Date(0, 1, 1).functions.toDtgAtStartOfDay(),
+            mk_Date(0, 1, 1).properties.dtgAtStartOfDay,
             Duration.fromDays(0).functions.toDtgAfterFirstDtg()
         )
         assertEquals(
-            mk_Date(0, 2, 7).functions.toDtgAtStartOfDay(),
+            mk_Date(0, 2, 7).properties.dtgAtStartOfDay,
             Duration.fromDays(37).functions.toDtgAfterFirstDtg()
         )
         assertEquals(
@@ -931,20 +1010,38 @@ class ISO8601Test {
     @Test
     fun durationToDtgAfterGivenDtgTest() {
         assertEquals(
-            mk_Date(1000, 1, 6).functions.toDtgAtStartOfDay(),
-            Duration.fromDays(5).functions.toDtgAfterGivenDtg(mk_Date(1000, 1, 1).functions.toDtgAtStartOfDay())
+            mk_Date(1000, 1, 6).properties.dtgAtStartOfDay,
+            Duration.fromDays(5).functions.toDtgAfterGivenDtg(
+                mk_Date(
+                    1000,
+                    1,
+                    1
+                ).properties.dtgAtStartOfDay
+            )
         )
         assertEquals(
-            mk_Date(100, 1, 1).functions.toDtgAtStartOfDay(),
-            Duration.fromDays(0).functions.toDtgAfterGivenDtg(mk_Date(100, 1, 1).functions.toDtgAtStartOfDay())
+            mk_Date(100, 1, 1).properties.dtgAtStartOfDay,
+            Duration.fromDays(0).functions.toDtgAfterGivenDtg(mk_Date(100, 1, 1).properties.dtgAtStartOfDay)
         )
         assertEquals(
-            mk_Date(2000, 3, 31).functions.toDtgAtStartOfDay(),
-            Duration.fromDays(90).functions.toDtgAfterGivenDtg(mk_Date(2000, 1, 1).functions.toDtgAtStartOfDay())
+            mk_Date(2000, 3, 31).properties.dtgAtStartOfDay,
+            Duration.fromDays(90).functions.toDtgAfterGivenDtg(
+                mk_Date(
+                    2000,
+                    1,
+                    1
+                ).properties.dtgAtStartOfDay
+            )
         )
         assertEquals(
             mk_Dtg(mk_Date(2024, 12, 31), mk_Time(13, 0, 0, 0)),
-            Duration.fromHours(13).functions.toDtgAfterGivenDtg(mk_Date(2024, 12, 31).functions.toDtgAtStartOfDay())
+            Duration.fromHours(13).functions.toDtgAfterGivenDtg(
+                mk_Date(
+                    2024,
+                    12,
+                    31
+                ).properties.dtgAtStartOfDay
+            )
         )
     }
 
@@ -952,16 +1049,16 @@ class ISO8601Test {
     fun dtgToDurationSinceFirstDtgTest() {
         assertEquals(
             Duration.fromDays(6),
-            mk_Date(0, 1, 7).functions.toDtgAtStartOfDay().functions.toDurationSinceFirstDtg()
+            mk_Date(0, 1, 7).properties.dtgAtStartOfDay.properties.durationSinceFirstDtg
         )
         assertEquals(
             Duration.fromDays(0),
-            mk_Date(0, 1, 1).functions.toDtgAtStartOfDay().functions.toDurationSinceFirstDtg()
+            mk_Date(0, 1, 1).properties.dtgAtStartOfDay.properties.durationSinceFirstDtg
         )
 
         assertEquals(
             Duration.fromDays(37),
-            mk_Date(0, 2, 7).functions.toDtgAtStartOfDay().functions.toDurationSinceFirstDtg()
+            mk_Date(0, 2, 7).properties.dtgAtStartOfDay.properties.durationSinceFirstDtg
         )
     }
 
@@ -985,8 +1082,8 @@ class ISO8601Test {
 
     @Test
     fun dateToDurationSinceFirstDateTest() {
-        assertEquals(Duration.fromDays(3), mk_Date(0, 1, 4).functions.toDurationSinceFirstDate())
-        assertEquals(Duration.fromDays(0), mk_Date(0, 1, 1).functions.toDurationSinceFirstDate())
+        assertEquals(Duration.fromDays(3), mk_Date(0, 1, 4).properties.durationSinceFirstDate)
+        assertEquals(Duration.fromDays(0), mk_Date(0, 1, 1).properties.durationSinceFirstDate)
     }
 
     @Test
@@ -1010,8 +1107,8 @@ class ISO8601Test {
 
     @Test
     fun timeToDurationSinceFirstTimeTest() {
-        assertEquals(Duration.fromHours(4), mk_Time(4, 0, 0, 0).functions.toDurationSinceFirstTime())
-        assertEquals(Duration.fromHours(0), mk_Time(0, 0, 0, 0).functions.toDurationSinceFirstTime())
+        assertEquals(Duration.fromHours(4), mk_Time(4, 0, 0, 0).properties.durationSinceFirstTime)
+        assertEquals(Duration.fromHours(0), mk_Time(0, 0, 0, 0).properties.durationSinceFirstTime)
     }
 
     @Test
@@ -1021,35 +1118,35 @@ class ISO8601Test {
             mk_TimeInZone(
                 mk_Time(3, 0, 0, 0),
                 mk_Offset(Duration.fromHours(1), PlusOrMinus.Minus)
-            ).functions.toNormalisedDurationSinceFirstTime()
+            ).properties.normalisedDurationSinceFirstTime
         )
         assertEquals(
             Duration.fromHours(2),
             mk_TimeInZone(
                 mk_Time(3, 0, 0, 0),
                 mk_Offset(Duration.fromHours(1), PlusOrMinus.Plus)
-            ).functions.toNormalisedDurationSinceFirstTime()
+            ).properties.normalisedDurationSinceFirstTime
         )
         assertEquals(
             Duration.fromHours(1),
             mk_TimeInZone(
                 mk_Time(23, 0, 0, 0),
                 mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
-            ).functions.toNormalisedDurationSinceFirstTime()
+            ).properties.normalisedDurationSinceFirstTime
         )
         assertEquals(
             Duration.fromHours(23),
             mk_TimeInZone(
                 mk_Time(1, 0, 0, 0),
                 mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
-            ).functions.toNormalisedDurationSinceFirstTime()
+            ).properties.normalisedDurationSinceFirstTime
         )
         assertEquals(
             Duration.fromHours(0),
             mk_TimeInZone(
                 FirstTime,
                 mk_Offset(Duration.fromHours(0), PlusOrMinus.None)
-            ).functions.toNormalisedDurationSinceFirstTime()
+            ).properties.normalisedDurationSinceFirstTime
         )
     }
 
@@ -1058,20 +1155,20 @@ class ISO8601Test {
         assertEquals(
             Duration.fromDays(5),
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
-            ).functions.intervalDuration()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
+            ).properties.duration
         )
         assertEquals(
             Duration.fromMillis(1),
-            FirstDtg.functions.instant().functions.intervalDuration()
+            FirstDtg.properties.instant.properties.duration
         )
         assertEquals(
             Duration.fromHours(2),
             mk_Interval(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(3, 0, 0, 0))
-            ).functions.intervalDuration()
+            ).properties.duration
         )
     }
 
@@ -1082,40 +1179,32 @@ class ISO8601Test {
         }
         assertEquals(
             true,
-            mk_Dtg(mk_Date(0, 1, 1), mk_Time(10, 0, 0, 0)).functions.finestGranularity(
-                Duration.fromHours(1)
-            )
+            mk_Dtg(mk_Date(0, 1, 1), mk_Time(10, 0, 0, 0)).functions.finestGranularity(Duration.fromHours(1))
         )
         assertEquals(
             false,
-            mk_Dtg(mk_Date(0, 1, 1), mk_Time(10, 0, 0, 0)).functions.finestGranularity(
-                Duration.fromHours(3)
-            )
+            mk_Dtg(mk_Date(0, 1, 1), mk_Time(10, 0, 0, 0)).functions.finestGranularity(Duration.fromHours(3))
         )
     }
 
     @Test
     fun intervalFinestGranularityTest() {
         assertFailsWith<PreconditionFailure> {
-            FirstDtg.functions.instant().functions.finestGranularity(Duration.fromDays(0))
+            FirstDtg.properties.instant.functions.finestGranularity(Duration.fromDays(0))
         }
         assertEquals(
             true,
             mk_Interval(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(3, 0, 0, 0))
-            ).functions.finestGranularity(
-                Duration.fromHours(1)
-            )
+            ).functions.finestGranularity(Duration.fromHours(1))
         )
         assertEquals(
             false,
             mk_Interval(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(10, 0, 0, 0))
-            ).functions.finestGranularity(
-                Duration.fromHours(2)
-            )
+            ).functions.finestGranularity(Duration.fromHours(2))
         )
 
     }
@@ -1124,7 +1213,7 @@ class ISO8601Test {
     fun minDtgTest() {
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
-            minDtg(
+            DtgUtilities.minDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 6), mk_Time(1, 0, 0, 0)),
@@ -1135,7 +1224,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(0, 10, 0, 0)),
-            minDtg(
+            DtgUtilities.minDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(0, 10, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(14, 0, 0, 0)),
@@ -1146,7 +1235,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Dtg(mk_Date(0, 1, 1), mk_Time(1, 0, 0, 0)),
-            minDtg(
+            DtgUtilities.minDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(0, 1, 1), mk_Time(1, 0, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
@@ -1161,7 +1250,7 @@ class ISO8601Test {
     fun maxDtgTest() {
         assertEquals(
             mk_Dtg(mk_Date(1990, 2, 1), mk_Time(1, 0, 0, 0)),
-            maxDtg(
+            DtgUtilities.maxDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 6), mk_Time(1, 0, 0, 0)),
@@ -1172,7 +1261,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(14, 0, 0, 0)),
-            maxDtg(
+            DtgUtilities.maxDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(0, 10, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(14, 0, 0, 0)),
@@ -1183,7 +1272,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Dtg(mk_Date(9999, 1, 1), mk_Time(1, 0, 0, 0)),
-            maxDtg(
+            DtgUtilities.maxDtg(
                 mk_Set1(
                     mk_Dtg(mk_Date(0, 1, 1), mk_Time(1, 0, 0, 0)),
                     mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
@@ -1199,7 +1288,7 @@ class ISO8601Test {
     fun minDateTest() {
         assertEquals(
             mk_Date(1990, 1, 4),
-            minDate(
+            DateUtilities.minDate(
                 mk_Set1(
                     mk_Date(1990, 4, 1),
                     mk_Date(1990, 1, 4),
@@ -1210,7 +1299,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Date(0, 1, 1),
-            minDate(
+            DateUtilities.minDate(
                 mk_Set1(
                     mk_Date(0, 1, 1),
                     mk_Date(1990, 1, 1),
@@ -1225,7 +1314,7 @@ class ISO8601Test {
     fun maxDateTest() {
         assertEquals(
             mk_Date(1990, 12, 1),
-            maxDate(
+            DateUtilities.maxDate(
                 mk_Set1(
                     mk_Date(1990, 4, 1),
                     mk_Date(1990, 1, 4),
@@ -1236,7 +1325,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Date(3000, 1, 1),
-            maxDate(
+            DateUtilities.maxDate(
                 mk_Set1(
                     mk_Date(0, 1, 1),
                     mk_Date(1990, 1, 1),
@@ -1251,7 +1340,7 @@ class ISO8601Test {
     fun minTimeTest() {
         assertEquals(
             mk_Time(0, 0, 12, 1),
-            minTime(
+            TimeUtilities.minTime(
                 mk_Set1(
                     mk_Time(1, 0, 4, 1),
                     mk_Time(0, 20, 1, 4),
@@ -1262,7 +1351,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Time(0, 0, 0, 0),
-            minTime(
+            TimeUtilities.minTime(
                 mk_Set1(
                     mk_Time(0, 0, 0, 0),
                     mk_Time(23, 0, 1, 1),
@@ -1277,7 +1366,7 @@ class ISO8601Test {
     fun maxTimeTest() {
         assertEquals(
             mk_Time(3, 0, 1, 31),
-            maxTime(
+            TimeUtilities.maxTime(
                 mk_Set1(
                     mk_Time(1, 0, 4, 1),
                     mk_Time(0, 20, 1, 4),
@@ -1288,7 +1377,7 @@ class ISO8601Test {
         )
         assertEquals(
             mk_Time(23, 0, 1, 1),
-            maxTime(
+            TimeUtilities.maxTime(
                 mk_Set1(
                     mk_Time(0, 0, 1, 1),
                     mk_Time(23, 0, 1, 1),
@@ -1303,13 +1392,13 @@ class ISO8601Test {
     @Test
     fun minDurationTest() {
         assertEquals(
-            Duration.fromMinutes(40), minDuration(
+            Duration.fromMinutes(40), DurationUtiltites.minDuration(
                 mk_Set1(Duration.fromHours(3), Duration.fromMinutes(40), Duration.fromDays(2), Duration.fromHours(5))
             )
         )
         assertEquals(
             Duration.fromHours(0),
-            minDuration(
+            DurationUtiltites.minDuration(
                 mk_Set1(Duration.fromHours(0), Duration.fromMinutes(14), Duration.fromDays(2), Duration.fromHours(23))
             )
         )
@@ -1319,13 +1408,13 @@ class ISO8601Test {
     fun maxDurationTest() {
         assertEquals(
             Duration.fromDays(2),
-            maxDuration(
+            DurationUtiltites.maxDuration(
                 mk_Set1(Duration.fromHours(3), Duration.fromMinutes(40), Duration.fromDays(2), Duration.fromHours(5))
             )
         )
         assertEquals(
             Duration.fromDays(2000000000),
-            maxDuration(
+            DurationUtiltites.maxDuration(
                 mk_Set1(
                     Duration.fromHours(0),
                     Duration.fromMinutes(14),
@@ -1340,13 +1429,13 @@ class ISO8601Test {
     fun sumDurationTest() {
         assertEquals(
             mk_Duration(204000000),
-            sumDuration(
+            DurationUtiltites.sumDuration(
                 mk_Seq(Duration.fromHours(3), Duration.fromMinutes(40), Duration.fromDays(2), Duration.fromHours(5))
             )
         )
         assertEquals(
             mk_Duration(0),
-            sumDuration(mk_Seq(Duration.fromHours(0), Duration.fromMinutes(0), Duration.fromDays(0)))
+            DurationUtiltites.sumDuration(mk_Seq(Duration.fromHours(0), Duration.fromMinutes(0), Duration.fromDays(0)))
         )
     }
 
@@ -1357,40 +1446,67 @@ class ISO8601Test {
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)),
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 1))
             ),
-            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)).functions.instant()
+            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(1, 0, 0, 0)).properties.instant
         )
         assertEquals(
             mk_Interval(
                 mk_Dtg(mk_Date(1990, 1, 1), mk_Time(23, 59, 59, 999)),
-                mk_Date(1990, 1, 2).functions.toDtgAtStartOfDay()
+                mk_Date(1990, 1, 2).properties.dtgAtStartOfDay
             ),
-            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(23, 59, 59, 999)).functions.instant()
+            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(23, 59, 59, 999)).properties.instant
         )
     }
 
     @Test
     fun nextDateWithSameDayAsGivenDateTest() {
-        assertEquals(mk_Date(1990, 2, 1), nextDateWithSameDayAsGivenDate(mk_Date(1990, 1, 1)))
-        assertEquals(mk_Date(1990, 3, 31), nextDateWithSameDayAsGivenDate(mk_Date(1990, 1, 31)))
+        assertEquals(
+            mk_Date(1990, 2, 1),
+            DateUtilities.nextDateWithSameDayAsGivenDate(mk_Date(1990, 1, 1))
+        )
+        assertEquals(
+            mk_Date(1990, 3, 31),
+            DateUtilities.nextDateWithSameDayAsGivenDate(mk_Date(1990, 1, 31))
+        )
     }
 
     @Test
     fun nextDateWithSameDayAsGivenDayTest() {
-        assertEquals(mk_Date(0, 1, 4), nextDateWithSameDayAsGivenDay(mk_Date(0, 1, 1), 4))
-        assertEquals(mk_Date(1990, 3, 31), nextDateWithSameDayAsGivenDay(mk_Date(1990, 1, 31), 31))
-        assertEquals(mk_Date(1, 1, 14), nextDateWithSameDayAsGivenDay(mk_Date(0, 12, 15), 14))
+        assertEquals(
+            mk_Date(0, 1, 4),
+            DateUtilities.nextDateWithSameDayAsGivenDay(mk_Date(0, 1, 1), 4)
+        )
+        assertEquals(
+            mk_Date(1990, 3, 31),
+            DateUtilities.nextDateWithSameDayAsGivenDay(mk_Date(1990, 1, 31), 31)
+        )
+        assertEquals(
+            mk_Date(1, 1, 14),
+            DateUtilities.nextDateWithSameDayAsGivenDay(mk_Date(0, 12, 15), 14)
+        )
     }
 
     @Test
     fun previousDateWithSameDayAsGivenDateTest() {
-        assertEquals(mk_Date(1990, 1, 3), previousDateWithSameDayAsGivenDate(mk_Date(1990, 2, 3)))
-        assertEquals(mk_Date(1989, 12, 3), previousDateWithSameDayAsGivenDate(mk_Date(1990, 1, 3)))
+        assertEquals(
+            mk_Date(1990, 1, 3),
+            DateUtilities.previousDateWithSameDayAsGivenDate(mk_Date(1990, 2, 3))
+        )
+        assertEquals(
+            mk_Date(1989, 12, 3),
+            DateUtilities.previousDateWithSameDayAsGivenDate(mk_Date(1990, 1, 3))
+        )
     }
 
     @Test
     fun previousDateWithSameDayAsGivenDayTest() {
-        assertEquals(mk_Date(1990, 1, 12), previousDateWithSameDayAsGivenDay(mk_Date(1990, 1, 31), 12))
-        assertEquals(mk_Date(1989, 12, 12), previousDateWithSameDayAsGivenDay(mk_Date(1990, 1, 4), 12))
+        assertEquals(
+            mk_Date(1990, 1, 12),
+            DateUtilities.previousDateWithSameDayAsGivenDay(mk_Date(1990, 1, 31), 12)
+        )
+        assertEquals(
+            mk_Date(1989, 12, 12),
+            DateUtilities.previousDateWithSameDayAsGivenDay(mk_Date(1990, 1, 4), 12)
+        )
     }
 
     @Test
@@ -1402,7 +1518,7 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(5, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
                 )
-            ).functions.normalise()
+            ).properties.normalised
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(7, 0, 0, 0)),
@@ -1411,7 +1527,7 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(5, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
                 )
-            ).functions.normalise()
+            ).properties.normalised
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(5, 0, 0, 0)),
@@ -1420,7 +1536,7 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(5, 0, 0, 0), mk_Offset(Duration.fromHours(0), PlusOrMinus.None)
                 )
-            ).functions.normalise()
+            ).properties.normalised
         )
     }
 
@@ -1430,31 +1546,31 @@ class ISO8601Test {
             mk_NormalisedTime(mk_Time(7, 23, 12, 0), PlusOrMinus.None),
             mk_TimeInZone(
                 mk_Time(5, 23, 12, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
-            ).functions.normaliseTimeInZone()
+            ).properties.normalisedTime
         )
         assertEquals(
             mk_NormalisedTime(mk_Time(1, 23, 12, 0), PlusOrMinus.Minus),
             mk_TimeInZone(
                 mk_Time(23, 23, 12, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
-            ).functions.normaliseTimeInZone()
+            ).properties.normalisedTime
         )
         assertEquals(
             mk_NormalisedTime(mk_Time(23, 23, 12, 0), PlusOrMinus.None),
             mk_TimeInZone(
                 mk_Time(23, 23, 12, 0), mk_Offset(Duration.fromHours(0), PlusOrMinus.None)
-            ).functions.normaliseTimeInZone()
+            ).properties.normalisedTime
         )
         assertEquals(
             mk_NormalisedTime(mk_Time(3, 23, 12, 0), PlusOrMinus.None),
             mk_TimeInZone(
                 mk_Time(5, 23, 12, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
-            ).functions.normaliseTimeInZone()
+            ).properties.normalisedTime
         )
         assertEquals(
             mk_NormalisedTime(mk_Time(23, 23, 12, 0), PlusOrMinus.Plus),
             mk_TimeInZone(
                 mk_Time(1, 23, 12, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
-            ).functions.normaliseTimeInZone()
+            ).properties.normalisedTime
         )
 
     }
@@ -1463,19 +1579,19 @@ class ISO8601Test {
     fun formatDtgTest() {
         assertEquals(
             "1990-01-01T03:00:00",
-            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(3, 0, 0, 0)).functions.format()
+            mk_Dtg(mk_Date(1990, 1, 1), mk_Time(3, 0, 0, 0)).properties.formatted
         )
         assertEquals(
             "1990-10-11T03:00:01",
-            mk_Dtg(mk_Date(1990, 10, 11), mk_Time(3, 0, 1, 0)).functions.format()
+            mk_Dtg(mk_Date(1990, 10, 11), mk_Time(3, 0, 1, 0)).properties.formatted
         )
         assertEquals(
             "0000-01-01T03:00:00",
-            mk_Dtg(mk_Date(0, 1, 1), mk_Time(3, 0, 0, 0)).functions.format()
+            mk_Dtg(mk_Date(0, 1, 1), mk_Time(3, 0, 0, 0)).properties.formatted
         )
         assertEquals(
             "9999-01-01T03:00:00.010",
-            mk_Dtg(mk_Date(9999, 1, 1), mk_Time(3, 0, 0, 10)).functions.format()
+            mk_Dtg(mk_Date(9999, 1, 1), mk_Time(3, 0, 0, 10)).properties.formatted
         )
     }
 
@@ -1488,7 +1604,7 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
                 )
-            ).functions.format()
+            ).properties.formatted
         )
         assertEquals(
             "1990-01-01T03:00:00-02:00",
@@ -1497,7 +1613,7 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
                 )
-            ).functions.format()
+            ).properties.formatted
         )
         assertEquals(
             "1990-01-01T03:00:00Z",
@@ -1506,21 +1622,21 @@ class ISO8601Test {
                 mk_TimeInZone(
                     mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(0), PlusOrMinus.None)
                 )
-            ).functions.format()
+            ).properties.formatted
         )
     }
 
     @Test
     fun formatDateTest() {
-        assertEquals("1990-01-01", mk_Date(1990, 1, 1).functions.format())
-        assertEquals("0000-01-01", mk_Date(0, 1, 1).functions.format())
+        assertEquals("1990-01-01", mk_Date(1990, 1, 1).properties.formatted)
+        assertEquals("0000-01-01", mk_Date(0, 1, 1).properties.formatted)
     }
 
     @Test
     fun formatTimeTest() {
-        assertEquals("10:07:14", mk_Time(10, 7, 14, 0).functions.format())
-        assertEquals("00:00:00", mk_Time(0, 0, 0, 0).functions.format())
-        assertEquals("00:00:00.050", mk_Time(0, 0, 0, 50).functions.format())
+        assertEquals("10:07:14", mk_Time(10, 7, 14, 0).properties.formatted)
+        assertEquals("00:00:00", mk_Time(0, 0, 0, 0).properties.formatted)
+        assertEquals("00:00:00.050", mk_Time(0, 0, 0, 50).properties.formatted)
     }
 
     @Test
@@ -1529,27 +1645,33 @@ class ISO8601Test {
             "03:00:00+02:00",
             mk_TimeInZone(
                 mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus)
-            ).functions.format()
+            ).properties.formatted
         )
         assertEquals(
             "03:00:00-02:00",
             mk_TimeInZone(
                 mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(2), PlusOrMinus.Minus)
-            ).functions.format()
+            ).properties.formatted
         )
         assertEquals(
             "03:00:00Z",
             mk_TimeInZone(
                 mk_Time(3, 0, 0, 0), mk_Offset(Duration.fromHours(0), PlusOrMinus.None)
-            ).functions.format()
+            ).properties.formatted
         )
     }
 
     @Test
     fun formatOffsetTest() {
-        assertEquals("+02:00", mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus).functions.format())
-        assertEquals("-03:00", mk_Offset(Duration.fromHours(3), PlusOrMinus.Minus).functions.format())
-        assertEquals("00:00", mk_Offset(NoDuration, PlusOrMinus.None).functions.format())
+        assertEquals(
+            "+02:00",
+            mk_Offset(Duration.fromHours(2), PlusOrMinus.Plus).properties.formatted
+        )
+        assertEquals(
+            "-03:00",
+            mk_Offset(Duration.fromHours(3), PlusOrMinus.Minus).properties.formatted
+        )
+        assertEquals("00:00", mk_Offset(NoDuration, PlusOrMinus.None).properties.formatted)
     }
 
     @Test
@@ -1557,42 +1679,48 @@ class ISO8601Test {
         assertEquals(
             "1990-01-01T00:00:00/1990-01-06T00:00:00",
             mk_Interval(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 6).functions.toDtgAtStartOfDay()
-            ).functions.format()
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 6).properties.dtgAtStartOfDay
+            ).properties.formatted
         )
         assertEquals(
             "1990-10-10T00:00:00/1991-01-06T05:00:00",
             mk_Interval(
                 mk_Dtg(mk_Date(1990, 10, 10), mk_Time(0, 0, 0, 0)),
                 mk_Dtg(mk_Date(1991, 1, 6), mk_Time(5, 0, 0, 0))
-            ).functions.format()
+            ).properties.formatted
         )
     }
 
     @Test
     fun formatDurationTest() {
-        assertEquals("P2DT6H", Duration.fromHours(6).functions.addDuration(Duration.fromDays(2)).functions.format())
-        assertEquals("PT0S", Duration.fromHours(0).functions.format())
+        assertEquals(
+            "P2DT6H",
+            Duration.fromHours(6).functions.addDuration(Duration.fromDays(2)).properties.formatted
+        )
+        assertEquals("PT0S", Duration.fromHours(0).properties.formatted)
         assertEquals(
             "PT1.001S",
-            Duration.fromSeconds(1).functions.addDuration(Duration.fromMillis(1)).functions.format()
+            Duration.fromSeconds(1).functions.addDuration(Duration.fromMillis(1)).properties.formatted
         )
     }
 
     @Test
     fun dtgAddMonthsTest() {
         assertEquals(
-            mk_Date(1990, 3, 31).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 31).functions.toDtgAtStartOfDay().functions.addMonths(2)
+            mk_Date(1990, 3, 31).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 31).properties.dtgAtStartOfDay.functions.addMonths(2)
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 3, 28), mk_Time(3, 0, 0, 0)),
-            mk_Dtg(mk_Date(1990, 1, 31), mk_Time(3, 0, 0, 0)).functions.addMonths(1).functions.addMonths(1)
+            mk_Dtg(
+                mk_Date(1990, 1, 31),
+                mk_Time(3, 0, 0, 0)
+            ).functions.addMonths(1).functions.addMonths(1)
         )
         assertEquals(
-            mk_Date(1991, 1, 30).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 11, 30).functions.toDtgAtStartOfDay().functions.addMonths(2)
+            mk_Date(1991, 1, 30).properties.dtgAtStartOfDay,
+            mk_Date(1990, 11, 30).properties.dtgAtStartOfDay.functions.addMonths(2)
         )
     }
 
@@ -1600,16 +1728,19 @@ class ISO8601Test {
     fun dtgSubtractMonthsTest() {
         assertFailsWith<PreconditionFailure> { FirstDtg.functions.subtractMonths(1) }
         assertEquals(
-            mk_Date(1990, 2, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 4, 2).functions.toDtgAtStartOfDay().functions.subtractMonths(2)
+            mk_Date(1990, 2, 2).properties.dtgAtStartOfDay,
+            mk_Date(1990, 4, 2).properties.dtgAtStartOfDay.functions.subtractMonths(2)
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 28), mk_Time(3, 0, 0, 0)),
-            (mk_Dtg(mk_Date(1990, 3, 31), mk_Time(3, 0, 0, 0))).functions.subtractMonths(1).functions.subtractMonths(1)
+            (mk_Dtg(
+                mk_Date(1990, 3, 31),
+                mk_Time(3, 0, 0, 0)
+            )).functions.subtractMonths(1).functions.subtractMonths(1)
         )
         assertEquals(
-            mk_Date(1990, 11, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1991, 1, 2).functions.toDtgAtStartOfDay().functions.subtractMonths(2)
+            mk_Date(1990, 11, 2).properties.dtgAtStartOfDay,
+            mk_Date(1991, 1, 2).properties.dtgAtStartOfDay.functions.subtractMonths(2)
         )
     }
 
@@ -1622,7 +1753,8 @@ class ISO8601Test {
             mk_Date(1990, 3, 31), mk_Date(1990, 1, 31).functions.addMonths(2)
         )
         assertEquals(
-            mk_Date(1990, 3, 28), mk_Date(1990, 1, 31).functions.addMonths(1).functions.addMonths(1)
+            mk_Date(1990, 3, 28),
+            mk_Date(1990, 1, 31).functions.addMonths(1).functions.addMonths(1)
         )
         assertEquals(
             mk_Date(1991, 1, 30), mk_Date(1990, 11, 30).functions.addMonths(2)
@@ -1639,7 +1771,8 @@ class ISO8601Test {
             mk_Date(1990, 2, 2), mk_Date(1990, 4, 2).functions.subtractMonths(2)
         )
         assertEquals(
-            mk_Date(1990, 1, 28), mk_Date(1990, 3, 31).functions.subtractMonths(1).functions.subtractMonths(1)
+            mk_Date(1990, 1, 28),
+            mk_Date(1990, 3, 31).functions.subtractMonths(1).functions.subtractMonths(1)
         )
         assertEquals(
             mk_Date(1990, 11, 2), mk_Date(1991, 1, 2).functions.subtractMonths(2)
@@ -1649,28 +1782,28 @@ class ISO8601Test {
     @Test
     fun dtgAddDays() {
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.addDays(0)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.addDays(0)
         )
         assertEquals(
-            mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.addDays(9)
+            mk_Date(1990, 1, 10).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.addDays(9)
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 2, 1), LastTime),
             mk_Dtg(mk_Date(1990, 1, 1), LastTime).functions.addDays(31)
         )
         assertEquals(
-            mk_Date(1990, 2, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.addDays(32)
+            mk_Date(1990, 2, 2).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.addDays(32)
         )
         assertEquals(
-            mk_Date(1992, 3, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1992, 2, 1).functions.toDtgAtStartOfDay().functions.addDays(29)
+            mk_Date(1992, 3, 1).properties.dtgAtStartOfDay,
+            mk_Date(1992, 2, 1).properties.dtgAtStartOfDay.functions.addDays(29)
         )
         assertEquals(
-            mk_Date(1991, 1, 2).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.addDays(366)
+            mk_Date(1991, 1, 2).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.addDays(366)
         )
     }
 
@@ -1678,28 +1811,28 @@ class ISO8601Test {
     fun dtgSubtractDays() {
         assertFailsWith<PreconditionFailure> { FirstDtg.functions.subtractDays(5) }
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay().functions.subtractDays(0)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay.functions.subtractDays(0)
         )
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 1, 10).functions.toDtgAtStartOfDay().functions.subtractDays(9)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1990, 1, 10).properties.dtgAtStartOfDay.functions.subtractDays(9)
         )
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 2, 1).functions.toDtgAtStartOfDay().functions.subtractDays(31)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1990, 2, 1).properties.dtgAtStartOfDay.functions.subtractDays(31)
         )
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1990, 2, 2).functions.toDtgAtStartOfDay().functions.subtractDays(32)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1990, 2, 2).properties.dtgAtStartOfDay.functions.subtractDays(32)
         )
         assertEquals(
-            mk_Date(1992, 2, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1992, 3, 1).functions.toDtgAtStartOfDay().functions.subtractDays(29)
+            mk_Date(1992, 2, 1).properties.dtgAtStartOfDay,
+            mk_Date(1992, 3, 1).properties.dtgAtStartOfDay.functions.subtractDays(29)
         )
         assertEquals(
-            mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-            mk_Date(1991, 1, 2).functions.toDtgAtStartOfDay().functions.subtractDays(366)
+            mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+            mk_Date(1991, 1, 2).properties.dtgAtStartOfDay.functions.subtractDays(366)
         )
     }
 
@@ -1716,101 +1849,122 @@ class ISO8601Test {
     @Test
     fun dateSubtractDays() {
         assertFailsWith<PreconditionFailure> { FirstDate.functions.subtractDays(5) }
-        assertEquals(mk_Date(1990, 1, 1), mk_Date(1990, 1, 1).functions.subtractDays(0))
-        assertEquals(mk_Date(1990, 1, 1), mk_Date(1990, 1, 10).functions.subtractDays(9))
-        assertEquals(mk_Date(1990, 1, 1), mk_Date(1990, 2, 1).functions.subtractDays(31))
-        assertEquals(mk_Date(1990, 1, 1), mk_Date(1990, 2, 2).functions.subtractDays(32))
-        assertEquals(mk_Date(1992, 2, 1), mk_Date(1992, 3, 1).functions.subtractDays(29))
-        assertEquals(mk_Date(1990, 1, 1), mk_Date(1991, 1, 2).functions.subtractDays(366))
+        assertEquals(
+            mk_Date(1990, 1, 1),
+            mk_Date(1990, 1, 1).functions.subtractDays(0)
+        )
+        assertEquals(
+            mk_Date(1990, 1, 1),
+            mk_Date(1990, 1, 10).functions.subtractDays(9)
+        )
+        assertEquals(
+            mk_Date(1990, 1, 1),
+            mk_Date(1990, 2, 1).functions.subtractDays(31)
+        )
+        assertEquals(
+            mk_Date(1990, 1, 1),
+            mk_Date(1990, 2, 2).functions.subtractDays(32)
+        )
+        assertEquals(
+            mk_Date(1992, 2, 1),
+            mk_Date(1992, 3, 1).functions.subtractDays(29)
+        )
+        assertEquals(
+            mk_Date(1990, 1, 1),
+            mk_Date(1991, 1, 2).functions.subtractDays(366)
+        )
     }
 
     @Test
     fun monthsBetweenDtgsTest() {
-        assertFailsWith<PreconditionFailure> { monthsBetweenDtgs(LastDtg, FirstDtg) }
+        assertFailsWith<PreconditionFailure> { DtgUtilities.monthsBetweenDtgs(LastDtg, FirstDtg) }
         assertEquals(
             0,
-            monthsBetweenDtgs(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(), mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            DtgUtilities.monthsBetweenDtgs(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             0,
-            monthsBetweenDtgs(
-                mk_Date(1990, 1, 12).functions.toDtgAtStartOfDay(), mk_Date(1990, 2, 1).functions.toDtgAtStartOfDay()
+            DtgUtilities.monthsBetweenDtgs(
+                mk_Date(1990, 1, 12).properties.dtgAtStartOfDay,
+                mk_Date(1990, 2, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             6,
-            monthsBetweenDtgs(
-                mk_Date(1990, 12, 12).functions.toDtgAtStartOfDay(), mk_Date(1991, 6, 13).functions.toDtgAtStartOfDay()
+            DtgUtilities.monthsBetweenDtgs(
+                mk_Date(1990, 12, 12).properties.dtgAtStartOfDay,
+                mk_Date(1991, 6, 13).properties.dtgAtStartOfDay
             )
         )
     }
 
     @Test
     fun yearsBetweenDtgsTest() {
-        assertFailsWith<PreconditionFailure> { yearsBetweenDtgs(LastDtg, FirstDtg) }
+        assertFailsWith<PreconditionFailure> { DtgUtilities.yearsBetweenDtgs(LastDtg, FirstDtg) }
         assertEquals(
             0,
-            yearsBetweenDtgs(
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 1, 1).functions.toDtgAtStartOfDay()
+            DtgUtilities.yearsBetweenDtgs(
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay,
+                mk_Date(1990, 1, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             0,
-            yearsBetweenDtgs(
-                mk_Date(1990, 1, 12).functions.toDtgAtStartOfDay(),
-                mk_Date(1990, 12, 1).functions.toDtgAtStartOfDay()
+            DtgUtilities.yearsBetweenDtgs(
+                mk_Date(1990, 1, 12).properties.dtgAtStartOfDay,
+                mk_Date(1990, 12, 1).properties.dtgAtStartOfDay
             )
         )
         assertEquals(
             2,
-            yearsBetweenDtgs(
-                mk_Date(1990, 1, 12).functions.toDtgAtStartOfDay(),
-                mk_Date(1992, 3, 13).functions.toDtgAtStartOfDay()
+            DtgUtilities.yearsBetweenDtgs(
+                mk_Date(1990, 1, 12).properties.dtgAtStartOfDay,
+                mk_Date(1992, 3, 13).properties.dtgAtStartOfDay
             )
         )
     }
 
     @Test
     fun isStringADateTest() {
-        assertEquals(true, isStringIsoDate("2018-04-01"))
-        assertEquals(false, isStringIsoDate("2018/04/01"))
-        assertEquals(false, isStringIsoDate("2018-04"))
-        assertEquals(false, isStringIsoDate("2018-02-30"))
-        assertEquals(false, isStringIsoDate("2018-AA-01"))
-        assertEquals(false, isStringIsoDate("20.8-04-01"))
-        assertEquals(false, isStringIsoDate("-128-04-01"))
-        assertEquals(false, isStringIsoDate("2018-24-01"))
-        assertEquals(false, isStringIsoDate("2018-04-66"))
+        assertEquals(true, DateFormattingUtilities.isStringIsoDate("2018-04-01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018/04/01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018-04"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018-02-30"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018-AA-01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("20.8-04-01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("-128-04-01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018-24-01"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDate("2018-04-66"))
 
     }
 
     @Test
     fun stringToDateTest() {
-        assertEquals(mk_Date(2018, 4, 1), stringToDate("2018-04-01"))
-        assertFailsWith<PreconditionFailure> { stringToDate("2018-04-41") }
+        assertEquals(mk_Date(2018, 4, 1), DateFormattingUtilities.stringToDate("2018-04-01"))
+        assertFailsWith<PreconditionFailure> { DateFormattingUtilities.stringToDate("2018-04-41") }
     }
 
     @Test
     fun isStringADtgTest() {
-        assertEquals(true, isStringIsoDtg("1990-01-01T00:00:00"))
-        assertEquals(false, isStringIsoDtg("-990-01-01T00:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-01!00:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T"))
-        assertEquals(false, isStringIsoDtg("1990-G1-01T00:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-13-01T00:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-41T00:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T24:00:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T00:60:00"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T00:00:60"))
-        assertEquals(false, isStringIsoDtg("19.0-01-01T00:00:00"))
+        assertEquals(true, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("-990-01-01T00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01!00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-G1-01T00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-13-01T00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-41T00:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T24:00:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:60:00"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:60"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("19.0-01-01T00:00:00"))
 
-        assertEquals(true, isStringIsoDtg("1990-01-01T00:00:00.000"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T00:00:00.FFF"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T00:00:00-000"))
-        assertEquals(false, isStringIsoDtg("1990-01-01T00:00:00.-01"))
+        assertEquals(true, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:00.000"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:00.FFF"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:00-000"))
+        assertEquals(false, DateFormattingUtilities.isStringIsoDtg("1990-01-01T00:00:00.-01"))
 
     }
 
@@ -1818,13 +1972,12 @@ class ISO8601Test {
     fun stringToDtgTest() {
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(12, 23, 0, 0)),
-            stringToDtg("1990-01-01T12:23:00")
+            DateFormattingUtilities.stringToDtg("1990-01-01T12:23:00")
         )
         assertEquals(
             mk_Dtg(mk_Date(1990, 1, 1), mk_Time(12, 23, 0, 1)),
-            stringToDtg("1990-01-01T12:23:00.001")
+            DateFormattingUtilities.stringToDtg("1990-01-01T12:23:00.001")
         )
-        assertFailsWith<PreconditionFailure> { stringToDtg("1990-01-01T25:24:00.001") }
+        assertFailsWith<PreconditionFailure> { DateFormattingUtilities.stringToDtg("1990-01-01T25:24:00.001") }
     }
 }
-
